requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
werkzeug==3.0.1
requests==2.31.0
numpy==1.26.2